requirements.txt
Django==1.1.1
south==0.6.2
django-authority==0.4
